added new traits
