+++ src/linear_algebra/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod length;
+pub(crate) mod matrix;
+pub(crate) mod matrix_product;
+pub(crate) mod random;
+pub(crate) mod size;
